+++ embedded-hal-bus/src/spi/refcell.rs
@@ -1,8 +1,6 @@
 use core::cell::RefCell;
 use embedded_hal::digital::OutputPin;
-use embedded_hal::spi::{
-    ErrorType, Operation, SpiBus, SpiBusRead, SpiBusWrite, SpiDevice, SpiDeviceRead, SpiDeviceWrite,
-};
+use embedded_hal::spi::{ErrorType, Operation, SpiBus, SpiDevice};
 
 use super::DeviceError;
 
@@ -32,66 +30,6 @@ where
     CS: OutputPin,
 {
     type Error = DeviceError<BUS::Error, CS::Error>;
-}
-
-impl<'a, Word: Copy + 'static, BUS, CS> SpiDeviceRead<Word> for RefCellDevice<'a, BUS, CS>
-where
-    BUS: SpiBusRead<Word>,
-    CS: OutputPin,
-{
-    fn read_transaction(&mut self, operations: &mut [&mut [Word]]) -> Result<(), Self::Error> {
-        let bus = &mut *self.bus.borrow_mut();
-
-        self.cs.set_low().map_err(DeviceError::Cs)?;
-
-        let mut op_res = Ok(());
-        for buf in operations {
-            if let Err(e) = bus.read(buf) {
-                op_res = Err(e);
-                break;
-            }
-        }
-
-        // On failure, it's important to still flush and deassert CS.
-        let flush_res = bus.flush();
-        let cs_res = self.cs.set_high();
-
-        op_res.map_err(DeviceError::Spi)?;
-        flush_res.map_err(DeviceError::Spi)?;
-        cs_res.map_err(DeviceError::Cs)?;
-
-        Ok(())
-    }
-}
-
-impl<'a, Word: Copy + 'static, BUS, CS> SpiDeviceWrite<Word> for RefCellDevice<'a, BUS, CS>
-where
-    BUS: SpiBusWrite<Word>,
-    CS: OutputPin,
-{
-    fn write_transaction(&mut self, operations: &[&[Word]]) -> Result<(), Self::Error> {
-        let bus = &mut *self.bus.borrow_mut();
-
-        self.cs.set_low().map_err(DeviceError::Cs)?;
-
-        let mut op_res = Ok(());
-        for buf in operations {
-            if let Err(e) = bus.write(buf) {
-                op_res = Err(e);
-                break;
-            }
-        }
-
-        // On failure, it's important to still flush and deassert CS.
-        let flush_res = bus.flush();
-        let cs_res = self.cs.set_high();
-
-        op_res.map_err(DeviceError::Spi)?;
-        flush_res.map_err(DeviceError::Spi)?;
-        cs_res.map_err(DeviceError::Cs)?;
-
-        Ok(())
-    }
 }
 
 impl<'a, Word: Copy + 'static, BUS, CS> SpiDevice<Word> for RefCellDevice<'a, BUS, CS>